requirements.txt
# Data Scraping
telethon==1.28.1
requests==2.31.0

# Data Cleaning and Transformation
pandas==2.1.1
numpy>=1.26.0
dbt-core>=1.6.0  # Updated to a newer version
dbt-postgres>=1.6.0  # Updated to match dbt-core
sqlalchemy==2.0.20

# Database
psycopg2-binary>=2.9.7

# Logging and Monitoring
loguru==0.7.0

# Utilities
python-dotenv==0.21.0

# Build Tools
wheel==0.37.1

# Dependency Resolutions
mashumaro>=3.9,<3.15  # Resolves conflict with dbt-adapters
Jinja2>=3.1.3,<4      # Resolves conflict with dbt-common